requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-sqlalchemy==3.0.5
flask-jwt-extended==4.5.3
flask-socketio==5.3.6
python-dotenv==1.0.0
openai==1.3.0
requests==2.31.0
websockets==11.0.3
qrcode==7.4.2
python-barcode==0.14.0
redis==5.0.0
celery==5.3.4
gunicorn==21.2.0 